requirements.txt
pika
docker
psycopg2-binary
